Introduced a proxy for the tenacity metrics stream
In a distributed environment the services may not be reachable from the
public Internet. Given a scenario where the microservice dashboard is
the only public available entry point, we must not call metrics/streams
on service machines directly from the javascript code. Instead we use a
proxy method in the microservice dashboard.